Commands

Description

This module contains implementations of instructions and expressions provided by language.

Documentation

(!??) :: Ord k => Map k a -> k -> Either k a

Find the value at a key. Returns 'Left k' when the element cannot be found.

getInventoryCount :: Action Int

Get player's inventory count.

assertNotNothing :: String -> Maybe a -> Action a

If 'Maybe a' is Nothing, terminate execution with a String message.

assertFound :: (String -> String) -> Either String a -> Action a

If 'Either String a' is Left, terminate execution with a message constructed with function and Left content.

showRoom :: Action Room

Show the description of a current room and descriptions of items inside it.

showAndExecuteOnEntryCurrentRoom :: Action ()

Show the description of a current room and descriptions of items inside it, then execute onEntry action.

noAction :: Action ()

Do nothing. Used for folding actions as neutral element.

showInventory :: Action ()

Show longNames of items in inventory.

getFinalMessage :: Action String

Get final message of a game.

callGlobalRoomCommandWithMessage :: String -> Name -> Action ()

Call a global room command. If specified command does not exist, terminate execution with message in String.

callCommandForCurrentRoom :: Name -> Action ()

Call a command in current room. If specified command does not exist, try to call global command.

callGlobalRoomCommand :: Name -> Action ()

Call a global room command. If specified command does not exist, terminate execution with an error.

callGlobalItemCommand :: Name -> ItemName -> Action ()

Call global item command, passing ItemName as the argument. If specified command does not exist, terminate execution with an error.

callGlobalOnEntry :: Action ()

Call global onEntry command.

takeItem :: ItemName -> Action ()

Take an item in current room into inventory.

giveItem :: ItemName -> Action ()

Give an item to inventory.

putItemInRoom :: ItemName -> Name -> Action ()

Put an item in a specified room.

putItem :: ItemName -> Action ()

Put an item in a current room.

dropItem :: ItemName -> Action ()

Drop an item from an inventory.

moveItemToRoom :: ItemName -> Name -> Action ()

Move an item from current room to a specified room.

removeItemFromRoom :: ItemName -> Name -> Action ()

Remove an item from specified room.

removeItem :: ItemName -> Action ()

Remove an item from current room.

discardItem :: ItemName -> Action ()

Discard an item (remove it from inventory and from current room).

dropAllItemsInCurrentRoom :: Action ()

Drop all items from inventory.

getCommandCount :: Action Int

Get a number of commands (also incorrect) already entered by a user

increaseCommandCount :: Action ()

Increase command count by one.

printMessage :: String -> Action ()

Print a String to a user, without new line at the end.

printMessageLine :: String -> Action ()

Print a String to a user, with a new line at the end.

printInt :: Int -> Action ()

Print a number to a user.

goToRoom :: Name -> Action ()

Go to specified room (changes currentRoomName).

changePlayerParameter :: Name -> (Int -> Int) -> Action ()

Change specified player parameter with a function.

getPlayerParameter :: Name -> Action Int

Get specified player parameter. If a parameter does not exist, terminate execution with an error.

changeEntityParameter :: Name -> Name -> (Int -> Int) -> Action ()

Change specified entity parameter with a function.

getEntityParameter :: Name -> Name -> Action Int

Get specified entity parameter. If a parameter does not exist, terminate execution with an error.

conditionallyPerformAction :: Action Bool -> Action () -> Action () -> Action ()

Perform either first or second action, regarding of the result of 'Action Bool'.

conditionallyEvaluateAction :: Action Bool -> (a -> Action ()) -> (a -> Action ()) -> a -> Action ()

The same as conditionallyPerformAction, but for parametrised actions.

checkIfItemIsInInventory :: ItemName -> Action Bool

Return true if item is in inventory.

checkIfInteractablePresent :: Name -> Action Bool

Return true if item is either in inventory or in current room.

checkIfRoomIsCurrent :: Name -> Action Bool

Return true if specified room is current room.

checkIfInteractablePresentInRoom :: Name -> Name -> Action Bool

Return true if specified item is in specified room.

saveGame :: Action ()

Prompt for file name and save the game state to a file.

restoreGame :: Action ()

Prompt for file name and restore the same state from a file.