DataStructures

Documentation

type Name = String

type ParamName = String

type Id = Int

type ItemName = Name

type Inventory = [ItemName]

type Hand = Maybe ItemName

type Visibility = Bool

type StIO a = StateT Game IO a

type Desc = Action String

type StParser a = StateT (Map Name Interactable, Action ()) Parser a

Stateful Parser with a map of interactables already read and default room entry action as a state.

(<??>) :: StParser a -> String -> StParser a

Name a stateful Parser. The lifted equivalent of <?> from attoparsec.

type Command = (Visibility, Name, Action ())

Representation of a command entered in a console by user

data Interactable

Constructors

Item

Fields

longName :: Name

getDescription :: Desc

getCommands :: [Command]

Entity

Fields

getDescription :: Desc

entityParameters :: Map Name Int

getCommands :: [Command]

Invalid

data Room

Constructors

Room

Fields

description :: Desc

onEntry :: Action ()

interactables :: [Name]

roomCommands :: [Command]

data Player

Structure that holds player parameters and items currently toting.

Constructors

Player

Fields

playerParameters :: Map Name Int

playerInventory :: Inventory

leftHand :: Hand

rightHand :: Hand

data Game

Structure that holds all information and state of the game.

Constructors

Game

Fields

player :: Player

initialMessage :: String

finalMessage :: Desc

globalOnEntry :: Action ()

globalRoomCommands :: [Command]

globalItemCommands :: ItemName -> [Command]

commandCount :: Int

rooms :: Map Name Room

currentRoomName :: Name

globalNameMap :: Map Name Interactable

newtype Action a

Encapsulates a 'StIO a' monad with an ability of immediate termination of a flow.

Constructors

Action

Fields

fromAction :: StIO (Either String a)

perform :: StIO a -> Action a

Lift StIO to an Action.

terminate :: String -> Action a

Terminate an action flow with a custom message in String.