Parsers.ArithmeticExpressionParser

Description

Provides an expressionParser parser for processing the arithmetic expressions. The possible operations are: +, -, *, / and % (mod) and are defined for Int. Moreover, the parser provides accessors to parameters (player.<parameter> and entity.<name>.<parameter>) and expressions: invcount (count of the player inventory), cmdcount (number of entered commands during the game) and rnd (random number).

Grammar

The arithmetic expression grammar in this module is defined as follows:

expr -> term expr'

expr' -> (+|-) term expr' | eps

term -> factor | term'

term' -> (*|/|%) factor term' | eps

factor -> invcount | cmdcount | rnd | number | player.<parameter> | entity.<name>.parameter | ( expr )

All operators are left-associative. Operators *,/,% have higher priority than + and -.

Parsing arithmetic expressions

commandCountParser :: Parser (Action Int)

Match a cmdcount expression.

inventoryCountParser :: Parser (Action Int)

Match an invcount expression.

randomNumberParser :: Parser (Action Int)

Match a rnd expression.

entityParameterAccessorParser :: Parser (Name, Name)

Match an expression of type entity.name.name.

playerParameterAccessorParser :: Parser Name

Match an expression of type player.name.

constantParser :: Parser (Action Int)

Match a number.

identificatorParser :: Parser (Action Int)

Match a parameter accessor expression.

mulOpParser :: Parser (Action Int -> Action Int -> Action Int)

Match multiplication or division operator.

addOpParser :: Parser (Action Int -> Action Int -> Action Int)

Match addition or subtraction operator.

termParser :: Parser (Action Int)

Match a single term of a sum.

termParser' :: Parser (Action Int -> Action Int)

Match a term' non-terminal from a grammar, preserving left-associativity of the operators.

expressionParser :: Parser (Action Int)

Match an arithmetic expression and return its symbolic representation in a parser.

expressionParser' :: Parser (Action Int -> Action Int)

Match a expr' non-terminal from a grammar, preserving left-associativity of the operators.

factorParser :: Parser (Action Int)

Match a single factor of a term.