Parsers.BooleanExpressionParser

Description

Provides a booleanExpressionParser parser for processing the boolean expressions. The possible operations are: && (logical and) and || (logical or).

Parser makes use of module ArithmeticExpressionParser to provide comparison operators (>,,=,<=,==,!=).

Moreover, the parser provides some useful tests: checking if object is in inventory (has <name>), if an object is present in current or other room (present <name> [in <room>]), or if player is in specified room (in <room>).

Expressions askYesNo and prompt <word> request an input from a user and check if, respectively, entered "yes"/"y" or <word>.

Grammar

The boolean expression grammar in this module is defined as follows:

exprB -> termB op exprB

op -> (&&|||)

termB -> askYesNo | present <name> | present <name> in <room> | in <room> | has <item> | true | false | expr comp expr | ( exprB )

comp -> > | < | >= | <= | == | !=

Logical operators are right-associative and have the same priority.

Parsing arithmetic expressions

askYesNoParser :: Parser (Action Bool)

Match an askYesNo prompt.

promptParser :: Parser (Action Bool)

Match a prompt word prompt.

hasItemParser :: Parser (Action Bool)

Match a has item test.

presentParser :: Parser (Action Bool)

Match a present item|entity test.

presentInRoomParser :: Parser (Action Bool)

Match a present item|entity in room test.

inRoomParser :: Parser (Action Bool)

Match an in room test.

comparisonOpParser :: Parser (Action Int -> Action Int -> Action Bool)

Match a comparison operator.

boolOpParser :: Parser (Action Bool -> Action Bool -> Action Bool)

Match either a boolean sum or product.

trueParser :: Parser (Action Bool)

Match true value. The following representations are correct: true, True, TRUE.

falseParser :: Parser (Action Bool)

Match false value. The following representations are correct: false, False, FALSE.

boolTermParser :: Parser (Action Bool)

Match a sigle term of a boolean expression.

booleanExpressionParser :: Parser (Action Bool)

Match a boolean expression.