Parsers.CodeParser

Description

Provides codeParser and itemCode parsers, which parse single block of command code, respectively, specific to Room and to Item.

Documentation

openCurlyBrace :: Parser Char

Match an opening curly brace ({) of code block.

closeCurlyBrace :: Parser Char

Match a closing curly brace (}) of code block.

printEmptyLineParser :: Parser (Action ())

Match a print. instruction.

printParser :: Parser (Action ())

Match a print <text> instruction.

printLineParser :: Parser (Action ())

Match a println <text> instruction.

printValueParser :: Parser (Action ())

Match a printval <expression> instruction.

gotoParser :: Parser (Action ())

Match a goto <room> instruction.

assignmentParser :: Parser (Action (Int -> Int))

Match an assignment instruction.

modificationParser :: Parser (Action (Int -> Int))

Match a modification instruction.

parameterFunctionParser :: Parser (Action (Int -> Int))

Match either an assignment or a modification instruction.

changePlayerParameterParser :: Parser (Action ())

Match an instruction of type player.<parameter> =+=-=/*= <expression>.

changeEntityParameterParser :: Parser (Action ())

Match an instruction of type entity.<entity name>.<parameter> =+=-=/*= <expression>.

conditionalParserF :: Parser (a -> Action ()) -> Parser (a -> Action ())

Match a conditional function instruction.

conditionalParser :: Parser (Action ()) -> Parser (Action ())

Match a conditional instruction.

takeSelfParser :: Parser (ItemName -> Action ())

Match a take instruction.

dropSelfParser :: Parser (ItemName -> Action ())

Match a drop instruction.

takeItemParser :: Parser (Action ())

Match a take <item name> instruction.

dropItemParser :: Parser (Action ())

Match a drop <item name> instruction.

dropAllParser :: Parser (Action ())

Match a dropall instruction.

discardItemParser :: Parser (ItemName -> Action ())

Match a discard instruction.

giveItemParser :: Parser (Action ())

Match a give <item name> instruction.

putItemParser :: Parser (Action ())

Match a put <item name> instruction.

removeItemParser :: Parser (Action ())

Match a remove <item name> instruction.

removeItemFromRoomParser :: Parser (Action ())

Match a remove <item> from <room> instruction.

putItemInRoomParser :: Parser (Action ())

Match a put <item> in <room> instruction.

moveItemToRoomParser :: Parser (Action ())

Match a move <item> to <room> instruction.

callForRoomParser :: Parser (Action ())

Match a call command instruction for a room.

quitParser :: Parser (Action ())

Match a quit or exit instruction.

commentParser :: Parser (Action ())

Match a comment.

saveParser :: Parser (Action ())

Match a save instruction.

restoreParser :: Parser (Action ())

Match a restore instruction.

showInventoryParser :: Parser (Action ())

Match a showInventory instruction.

lookParser :: Parser (Action ())

Match a look instruction.

globalOnEntryCallParser :: Parser (Action ())

Match a globalOnEntry instruction.

commonCodeLineParser :: Parser (Action ())

Match a single code line common for items and the rest.

codeLineParser :: Parser (Action ())

Match a single code line (or a conditional instruction)

codeParser :: Parser (Action ())

Match a code block with common instructions.

hasConditionalParser :: Parser (ItemName -> Action ()) -> Parser (ItemName -> Action ())

Match a special conditional instruction if has then ...

presentConditionalParser :: Parser (ItemName -> Action ()) -> Parser (ItemName -> Action ())

Match a special conditional instruction if present then ...

itemCodeLineParser :: Parser (ItemName -> Action ())

Match single code line with instructions specific to item.

itemCodeParser :: Parser (ItemName -> Action ())

Match a code block with instructions specific to item.

globalRoomParser :: Parser (Action ())

Match a global room.<command> instruction.

globalItemParser :: Parser (ItemName -> Action ())

Match a global item.<command> instruction.

globalExplicitItemParser :: Parser (Action ())

Match a global item.<command> <argument> instruction.