Parsers.Utilities

Documentation

isNewline :: Char -> Bool

Returns True if Char is a new line character (n or r).

skipSpaces :: Parser ()

Skip consecutive whitespace characters.

charWithSpaces :: Char -> Parser Char

Match a char with whitespace characters around.

stringWithSpaces :: Text -> Parser Text

Match a string with whitespace characters around.

newLines :: Parser ()

Match zero or more newline characters.

newLines1 :: Parser ()

Match one or more newline characters.

tabOrTwoSpaces :: Parser ()

Match a tab character (t) or two spaces.

tabs :: Int -> Parser ()

Match a number of tabs (or two spaces) specified by Int.

listParser :: Text -> Parser a -> Int -> String -> Parser [a]

Match a list of items stated by Text. 'Parser a' matches single item, Int is an indentation level and String is a message in case of parsing failure.

listParserSt :: Text -> StParser a -> Int -> String -> StParser [a]

Match a list of items stated by Text. Lifted version of listParser.

baseCodeLineParser :: Text -> Parser a -> String -> Parser a

Match a code line with keyword Text, instruction content parser 'Parser a'. String is a message in case of parsing failure.

multilineContentParser :: Text -> Int -> String -> Parser String

Match a multiline string, where Text is a keyword,Int is an indentation level and String is a message in case of parsing failure.

multilineContentParser' :: Int -> String -> Parser String

Match a multiline string, Int is an indentation level and String is a message in case of parsing failure.

liftPAction :: (a -> a -> b) -> Parser (Action a -> Action a -> Action b)

Lifts a two-argument function to Action and wraps it into a Parser.

liftPStIO :: (a -> a -> b) -> Parser (StIO a -> StIO a -> StIO b)

Lifts a two-argument function to StIO and wraps it into a Parser.

eitherResult' :: Result r -> Either String r

Version of attoparsec's function eitherResult, which in case of failure prints additionally a piece of unconsumed GAML code.